refactor: data와 model 레이어 분리
