+++ app/src/main/java/woowacourse/shopping/model/CartRepository.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.model
+
+interface CartRepository {
+    fun addCartProduct(product: Product)
+
+    fun getAllCartProducts(): List<Product>
+
+    fun deleteCartProduct(id: Int)
+}
