+++ app/src/main/java/woowacourse/shopping/model/ProductRepository.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.model
+
+interface ProductRepository {
+    fun getAllProducts(): List<Product>
+}
