+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -3,13 +3,13 @@ package woowacourse.shopping.ui
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.data.CartRepository
-import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.model.CartRepository
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.ProductRepository
 
 class MainViewModel(
-    private val productRepository: ProductRepository,
-    private val cartRepository: CartRepository,
+    private val defaultProductRepository: ProductRepository,
+    private val defaultCartRepository: CartRepository,
 ) : ViewModel() {
 
     private val _products: MutableLiveData<List<Product>> = MutableLiveData(emptyList())
@@ -18,13 +18,12 @@ class MainViewModel(
     private val _onProductAdded: MutableLiveData<Boolean> = MutableLiveData(false)
     val onProductAdded: LiveData<Boolean> get() = _onProductAdded
 
-
     fun addCartProduct(product: Product) {
-        cartRepository.addCartProduct(product)
+        defaultCartRepository.addCartProduct(product)
         _onProductAdded.value = true
     }
 
     fun getAllProducts() {
-        _products.value = productRepository.getAllProducts()
+        _products.value = defaultProductRepository.getAllProducts()
     }
 }
